+++ tests/Catalog/CatalogServiceTest.php
@@ -3097,7 +3097,7 @@ final class CatalogServiceTest extends ApiTestCase
 
         $product = $catalogService->getProductsByCode('6403500679952');
 
-        static::assertSame('', $product[0]->getImages()[0]->getAlt());
+        static::assertSame('', $product[0]->getImages()[0]->getAltText());
     }
 
     public function testGetProductWithImagesAlt(): void
@@ -3106,7 +3106,7 @@ final class CatalogServiceTest extends ApiTestCase
 
         $product = $catalogService->getProductById('38');
 
-        static::assertSame('', $product->getImages()[0]->getAlt());
+        static::assertSame('', $product->getImages()[0]->getAltText());
     }
 
     public function testSearchProductWithImagesAlt(): void
@@ -3116,7 +3116,7 @@ final class CatalogServiceTest extends ApiTestCase
         $result = $catalogService->search();
         $products = $result->getProducts();
 
-        static::assertSame('alt text', $products[0]->getMainImage()->getAlt());
+        static::assertSame('alt text', $products[0]->getMainImage()->getAltText());
     }
 
     public function testGetDeclinationByIDWithImagesAlt(): void
@@ -3136,6 +3136,6 @@ final class CatalogServiceTest extends ApiTestCase
         $product = $products[24];
 
         static::assertInstanceOf(Product::class, $product);
-        static::assertSame('', $product->getImages()[0]->getAlt());
+        static::assertSame('', $product->getImages()[0]->getAltText());
     }
 }